fixed file structure for events page.
