--- src/components/EventHolder/index.jsx
@@ -1,5 +1,0 @@
-import Holder from './holder'
-import Events from './events'
-
-export { Holder, Events }
-
